Release v0.9.0(beta)

+++ Sauron/BackEnd.pm
@@ -10,6 +10,7 @@ require Exporter;
 use NetAddr::IP; # For IPv6;
 use Sauron::DB;
 use Sauron::Util;
+use Sauron::SetupIO;
 use Sys::Syslog qw(:DEFAULT setlogsock);
 Sys::Syslog::setlogsock('unix');
 use Net::IP qw (:PROC);
@@ -178,7 +179,7 @@ sub write2log
   my $filename  = File::Basename::basename($0);
 
   Sys::Syslog::openlog($filename, "cons,pid", "debug");
-  Sys::Syslog::syslog("info", "$msg");
+  Sys::Syslog::syslog("info", encode_str("$msg"));
   Sys::Syslog::closelog();
 } # End of write2log
 
@@ -196,7 +197,7 @@ sub fix_bools($$) {
 }
 
 sub sauron_db_version() {
-  return "1.7"; # required db format version for this backend
+  return "1.8"; # required db format version for this backend
 }
 
 sub set_muser($) {
@@ -1577,6 +1578,13 @@ sub delete_zone($) {
 	       "WHERE h.zone=$id AND a.type=1 AND a.ref=h.id)");
   if ($res < 0) { db_rollback(); return -15; }
 
+  # srv_entries
+  print "<BR>Deleting SSHFP entries...\n";
+  $res=db_exec("DELETE FROM sshfp_entries WHERE id IN ( " .
+	       "SELECT a.id FROM sshfp_entries a, hosts h " .
+	       "WHERE h.zone=$id AND a.type=1 AND a.ref=h.id)");
+  if ($res < 0) { db_rollback(); return -15; }
+
   # arec_entries
   print "<BR>Deleting (sub)group entries...\n";
   $res=db_exec("DELETE FROM group_entries WHERE id IN ( " .
@@ -1925,13 +1933,15 @@ sub get_host($$) {
 
   get_array_field("ns_entries",3,"id,ns,comment","NS,Comments",
 		  "type=2 AND ref=$id ORDER BY ns",$rec,'ns_l');
+  get_array_field("ds_entries",6,"id,key_tag,algorithm,digest_type,digest,comment",
+                  "Key tag,Algorithm,Digest type,Digest,Comments",
+		  "type=2 AND ref=$id ORDER BY key_tag,algorithm,digest_type,digest",
+                  $rec,'ds_l');
   get_array_field("wks_entries",4,"id,proto,services,comment",
 		  "Proto,Services,Comments",
 		  "type=1 AND ref=$id ORDER BY proto,services",$rec,'wks_l');
   get_array_field("mx_entries",4,"id,pri,mx,comment","Priority,MX,Comments",
 		  "type=2 AND ref=$id ORDER BY pri,mx",$rec,'mx_l');
-  get_array_field("txt_entries",3,"id,txt,comment","TXT,Comments",
-		  "type=2 AND ref=$id ORDER BY id",$rec,'txt_l');
   get_array_field("dhcp_entries",3,"id,dhcp,comment","DHCP,Comments",
 		  "type=3 AND ref=$id ORDER BY id",$rec,'dhcp_l');
   get_array_field("dhcp_entries",3,"id,dhcp,comment","DHCP,Comments",
@@ -1939,9 +1949,17 @@ sub get_host($$) {
   get_array_field("printer_entries",3,"id,printer,comment","PRINTER,Comments",
 		  "type=2 AND ref=$id ORDER BY printer",$rec,'printer_l');
   get_array_field("srv_entries",6,"id,pri,weight,port,target,comment",
-		  "Priority,Weight,Port,Target",
+		  "Priority,Weight,Port,Target,Comments",
 		  "type=1 AND ref=$id ORDER BY port,pri,weight",$rec,'srv_l');
-
+  get_array_field("sshfp_entries",6,"id,algorithm,hashtype,fingerprint,comment",
+		  "Algorithm,Type,Fingerprint,Comments",
+		  "type=1 AND ref=$id ORDER BY algorithm,hashtype,fingerprint",$rec,'sshfp_l');
+  get_array_field("tlsa_entries",6,"id,usage,selector,matching_type,association_data,comment",
+		  "Usage,Selector,Matching Type,Asociation Data,Comments",
+		  "type=1 AND ref=$id ORDER BY usage,selector,matching_type,association_data",$rec,'tlsa_l');
+  get_array_field("txt_entries",3,"id,txt,comment",
+		  "Text,Comments",
+		  "type=2 AND ref=$id ORDER BY txt",$rec,'txt_l');
 # Get CNAME aliases.
   get_array_field("hosts",5,"0,id,domain,type,1","Domain,cname",
 	          "type=4 AND alias=$id ORDER BY domain",$rec,'alias_l');
@@ -2093,6 +2111,9 @@ sub update_host($) {
 
   $r=update_array_field("ns_entries",3,"ns,comment,type,ref",
 			'ns_l',$rec,"2,$id");
+  if ($r < 0) { db_rollback(); return -11; }
+  $r=update_array_field("ds_entries",6,"key_tag,algorithm,digest_type,digest,comment,type,ref",
+			'ds_l',$rec,"2,$id");
   if ($r < 0) { db_rollback(); return -12; }
   $r=update_array_field("wks_entries",4,"proto,services,comment,type,ref",
 			'wks_l',$rec,"1,$id");
@@ -2100,9 +2121,6 @@ sub update_host($) {
   $r=update_array_field("mx_entries",4,"pri,mx,comment,type,ref",
 			'mx_l',$rec,"2,$id");
   if ($r < 0) { db_rollback(); return -14; }
-  $r=update_array_field("txt_entries",3,"txt,comment,type,ref",
-			'txt_l',$rec,"2,$id");
-  if ($r < 0) { db_rollback(); return -15; }
   $r=update_array_field("dhcp_entries",3,"dhcp,comment,type,ref",
 			'dhcp_l',$rec,"3,$id");
   if ($r < 0) { db_rollback(); return -16; }
@@ -2113,24 +2131,35 @@ sub update_host($) {
 			"pri,weight,port,target,comment,type,ref",
 			'srv_l',$rec,"1,$id");
   if ($r < 0) { db_rollback(); return -18; }
-
+  $r=update_array_field("sshfp_entries",5,
+			"algorithm,hashtype,fingerprint,comment,type,ref",
+			'sshfp_l',$rec,"1,$id");
+  if ($r < 0) { db_rollback(); return -19; }
+  $r=update_array_field("tlsa_entries",6,
+			"usage,selector,matching_type,association_data,comment,type,ref",
+			'tlsa_l',$rec,"1,$id");
+  if ($r < 0) { db_rollback(); return -20; }
+  $r=update_array_field("txt_entries",3,
+			"txt,comment,type,ref",
+			'txt_l',$rec,"2,$id");
+  if ($r < 0) { db_rollback(); return -21; }
   $r=update_array_field("a_entries",4,"ip,reverse,forward,host",
 			'ip',$rec,"$id");
-  if ($r < 0) { db_rollback(); return -20; }
+  if ($r < 0) { db_rollback(); return -22; }
 
   if ($rec->{type}==7) {
     $r=update_array_field("arec_entries",2,"arec,host",
 			  'alias_a',$rec,"$id");
-    if ($r < 0) { db_rollback(); return -21; }
+    if ($r < 0) { db_rollback(); return -23; }
   }
 
   $r=update_array_field("group_entries",2,"grp,host",
 			'subgroups',$rec,"$id");
-  if ($r < 0) { db_rollback(); return -22; }
+  if ($r < 0) { db_rollback(); return -24; }
 
   $r=update_array_field("dhcp_entries",3,"dhcp,comment,type,ref",
 			'dhcp_l6',$rec,"13,$id");
-  if ($r < 0) { db_rollback(); return -23; }
+  if ($r < 0) { db_rollback(); return -25; }
 
   return db_commit();
 }
@@ -2208,11 +2237,23 @@ sub delete_host($) {
   $res=db_exec("DELETE FROM srv_entries WHERE type=1 AND ref=$id;");
   if ($res < 0) { db_rollback(); return -11; }
 
+  # sshfp_entries
+  $res=db_exec("DELETE FROM sshfp_entries WHERE type=1 AND ref=$id;");
+  if ($res < 0) { db_rollback(); return -15; }
+
+  # tlsa_entries
+  $res=db_exec("DELETE FROM tlsa_entries WHERE type=1 AND ref=$id;");
+  if ($res < 0) { db_rollback(); return -16; }
+
+  # txt_entries
+  $res=db_exec("DELETE FROM txt_entries WHERE type=2 AND ref=$id;");
+  if ($res < 0) { db_rollback(); return -17; }
+
   # group_entries
   $res=db_exec("DELETE FROM group_entries WHERE host=$id;");
   if ($res < 0) { db_rollback(); return -12; }
 
-  # Note that -13 and -14 are already in use!
+  # Note that -13, -14, -15 and -16 are already in use!
 
   $res=db_exec("DELETE FROM hosts WHERE id=$id;");
   if ($res < 0) { db_rollback(); return -50; }
@@ -2275,28 +2316,49 @@ sub add_host($) {
 			 'type,ref',"2,$id");
   if ($res < 0) { db_rollback(); return -4; }
 
+  # DSs (in Delegation form)
+  $res = add_array_field('ds_entries','key_tag,algorithm,digest_type,digest,comment','ds_l',
+                         $rec,
+			 'type,ref',"2,$id");
+  if ($res < 0) { db_rollback(); return -5; }
+
   # PRINTERs
   $res = add_array_field('printer_entries','printer,comment','printer_l',$rec,
 			 'type,ref',"2,$id");
-  if ($res < 0) { db_rollback(); return -5; }
+  if ($res < 0) { db_rollback(); return -6; }
 
   # SRVs
   $res = add_array_field('srv_entries','pri,weight,port,target,comment',
 			 'srv_l',$rec,'type,ref',"1,$id");
-  if ($res < 0) { db_rollback(); return -6; }
+  if ($res < 0) { db_rollback(); return -8; }
+
+  # SSHFPs
+  $res = add_array_field('sshfp_entries','algorithm,hashtype,fingerprint,comment',
+			 'sshfp_l',$rec,'type,ref',"1,$id");
+  if ($res < 0) { db_rollback(); return -9; }
+
+  # TLSAs
+  $res = add_array_field('tlsa_entries','usage,selector,matching_type,association_data,comment',
+			 'tlsa_l',$rec,'type,ref',"1,$id");
+  if ($res < 0) { db_rollback(); return -10; }
+
+  # TXTs
+  $res = add_array_field('txt_entries','txt,comment',
+			 'txt_l',$rec,'type,ref',"2,$id");
+  if ($res < 0) { db_rollback(); return -11; }
 
   # ARECs
   if ($rec->{type}==7) {
     $res=db_exec("INSERT INTO arec_entries (host,arec) VALUES($id,$a_id);");
-    if ($res < 0) { db_rollback(); return -7; }
+    if ($res < 0) { db_rollback(); return -12; }
   }
 
   # subgroups
   $res = add_array_field('group_entries','grp',
 			 'subgroups',$rec,'host',"$id");
-  if ($res < 0) { db_rollback(); return -8; }
+  if ($res < 0) { db_rollback(); return -13; }
 
-  return -10 if (db_commit() < 0);
+  return -20 if (db_commit() < 0);
   return $id;
 }
 
@@ -2305,8 +2367,9 @@ sub add_host($) {
 # TVu 02.04.2014.
 sub get_host_types() {
     return (0 => 'Any type', 1 => 'Host', 2 => 'Delegation', 3 => 'Plain MX',
-	    4 => 'Alias', 5 => 'Printer', 6 => 'Glue record', 7 => 'AREC Alias',
-	    8 => 'SRV record', 9 => 'DHCP only', 10 => 'Zone',
+	    4 => 'Alias', 5 => 'Printer', 6 => 'Glue', 7 => 'AREC Alias',
+	    8 => 'SRV', 9 => 'DHCP only', 10 => 'Zone', 11=>'SSHFP only',
+            12 => 'TLSA only', 13 => 'TXT',
 	    101 => 'Host reservation');
 }
 
@@ -3631,7 +3694,7 @@ sub get_key($$) {
 
   return -100 if (get_record("keys",
                       "type,ref,name,keytype,nametype,protocol,algorithm,".
-                      "mode,keysize,strength,publickey,secretkey,comments,".
+                      "mode,keysize,strength,publickey,secretkey,comment,".
 		      "cdate,cuser,mdate,muser", $id,$rec,"id"));
 
   add_std_fields($rec);
@@ -4012,12 +4075,17 @@ sub update_history($$$$$$) {
 }
 
 
-sub fix_utmp($) {
-  my($timeout) = @_;
+sub fix_utmp($$) {
+  my($timeout, $check_zero) = @_;
   my($i,$t,@q);
 
   $t=time - $timeout;
-  db_query("SELECT cookie,uid,sid FROM utmp WHERE last < $t;",\@q);
+  if ($check_zero) {
+    db_query("SELECT cookie,uid,sid FROM utmp WHERE last < $t and last != 0;",\@q);
+  } else {
+    db_query("SELECT cookie,uid,sid FROM utmp WHERE last < $t;",\@q);
+  }
+
   if (@q > 0) {
     for $i (0..$#q) {
       update_lastlog($q[$i][1],$q[$i][2],3,'','');
